Removed MultiObjective
Replaced it by over-writing the methods in Objective it self to sum
another function and multiply by scalar. The advantage is that the
result of the sum A + B is a copy of A with wrapped hessian, gradient
and value methods. It still retains all its properties (like overwritten
fit method, custom methods, etc).
Keep in mind that A + B + C will be a copy of A.

Still have to overwrite predicted and residuals in Misfit to return a
list of predicted and misfits, like MultiObjective did.