Copying the cache when adding Misfits
The cache was not being copied when added a Misfit to something else.
This meant that altering the original Misfit instance changed the cache
of the sum. That breaks computations in parallel because multiple places
will try to write to the same dict. Fixed this by shallow-copying the
cache when adding.

There is still a problem when adding 2 Misfits. The 'other' in the
addition (right side) is still the original. So this will also share a
cache. A way would be to copy the cache of 'other' if it has it.